+++ src/main/java/com/pekar/angelblock/tools/EnhancedHoe.java
@@ -10,6 +10,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
 public class EnhancedHoe extends ModHoe implements IModToolEnhanceable
@@ -58,10 +59,8 @@ public class EnhancedHoe extends ModHoe implements IModToolEnhanceable
             return;
 
         BlockState blockState = level.getBlockState(pos);
-        if (blockState.hasBlockEntity() || blockState != blockState.getBlock().defaultBlockState()) return;
-
-        float originHardness = blockState.getBlock().defaultDestroyTime();
-        if (originHardness == 0.0F) return;
+        var block = blockState.getBlock();
+        if (canNotBeMinedInGroup(blockState)) return;
 
         final int posX = pos.getX(), posY = pos.getY(), posZ = pos.getZ();
 
@@ -70,7 +69,7 @@ public class EnhancedHoe extends ModHoe implements IModToolEnhanceable
                 for (int z = posZ - 1; z <= posZ + 1; z++)
                 {
                     if (x == posX && y == posY && z == posZ) continue;
-                    onBlockMining(level, blockState, originHardness, new BlockPos(x, y, z), entityLiving);
+                    onBlockMining(level, new BlockPos(x, y, z), block, entityLiving);
                 }
     }
 
@@ -121,20 +120,27 @@ public class EnhancedHoe extends ModHoe implements IModToolEnhanceable
         return false;
     }
 
-    protected void onBlockMining(Level level, BlockState originBlockState, float originHardness, BlockPos pos, LivingEntity entityLiving)
+    protected void onBlockMining(Level level, BlockPos pos, Block originBlock, LivingEntity entityLiving)
     {
         var blockState = level.getBlockState(pos);
-
         var block = blockState.getBlock();
-        if (blockState.hasBlockEntity() || blockState != block.defaultBlockState()) return;
+
+        if (block != originBlock) return;
+        if (canNotBeMinedInGroup(blockState)) return;
 
         float hardness = block.defaultDestroyTime();
+        float originHardness = originBlock.defaultDestroyTime();
 
         if (hardness <= originHardness && isToolEffective(entityLiving, pos)
-                && (materialProperties.isSafeToBreak(entityLiving, pos) ||  entityLiving.isShiftKeyDown()))
+                && (materialProperties.isSafeToBreak(entityLiving, pos) || entityLiving.isShiftKeyDown()))
         {
-            if (utils.player.destroyBlockByMainHandTool(level, pos, entityLiving, blockState, block))
+            if (utils.player.destroyBlockByMainHandTool(level, pos, entityLiving, blockState))
                 damageMainHandItem(1, entityLiving);
         }
+    }
+
+    private boolean canNotBeMinedInGroup(BlockState blockState)
+    {
+        return blockState.hasBlockEntity();
     }
 }